renaming sinks and sources api to be consistent with the rest of Pulsar
