+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/admin/v3/Sink.java
@@ -19,7 +19,7 @@
 package org.apache.pulsar.broker.admin.v3;
 
 import io.swagger.annotations.Api;
-import org.apache.pulsar.broker.admin.impl.SinkBase;
+import org.apache.pulsar.broker.admin.impl.SinksBase;
 
 import javax.ws.rs.Consumes;
 import javax.ws.rs.Path;
@@ -30,5 +30,9 @@ import javax.ws.rs.core.MediaType;
 @Api(value = "/sink", description = "Sink admin apis", tags = "sink")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public class Sink extends SinkBase {
+@Deprecated
+/**
+ * @deprecated in favor of {@link Sinks}
+ */
+public class Sink extends SinksBase {
 }
