+++ pulsar-client-admin/src/main/java/org/apache/pulsar/client/admin/Sink.java
@@ -18,333 +18,10 @@
  */
 package org.apache.pulsar.client.admin;
 
-import org.apache.pulsar.client.admin.PulsarAdminException.NotAuthorizedException;
-import org.apache.pulsar.client.admin.PulsarAdminException.NotFoundException;
-import org.apache.pulsar.client.admin.PulsarAdminException.PreconditionFailedException;
-import org.apache.pulsar.common.functions.UpdateOptions;
-import org.apache.pulsar.common.io.ConnectorDefinition;
-import org.apache.pulsar.common.policies.data.SinkStatus;
-import org.apache.pulsar.common.io.SinkConfig;
-
-import java.util.List;
-
 /**
- * Admin interface for Sink management.
+ * @deprecated in favor of {@link Sinks}
  */
-public interface Sink {
-    /**
-     * Get the list of sinks.
-     * <p>
-     * Get the list of all the Pulsar Sinks.
-     * <p>
-     * Response Example:
-     *
-     * <pre>
-     * <code>["f1", "f2", "f3"]</code>
-     * </pre>
-     *
-     * @throws NotAuthorizedException
-     *             Don't have admin permission
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    List<String> listSinks(String tenant, String namespace) throws PulsarAdminException;
+@Deprecated
+public interface Sink extends Sinks {
 
-    /**
-     * Get the configuration for the specified sink.
-     * <p>
-     * Response Example:
-     *
-     * <pre>
-     * <code>{ serviceUrl : "http://my-broker.example.com:8080/" }</code>
-     * </pre>
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @return the sink configuration
-     *
-     * @throws NotAuthorizedException
-     *             You don't have admin permission to get the configuration of the cluster
-     * @throws NotFoundException
-     *             Cluster doesn't exist
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    SinkConfig getSink(String tenant, String namespace, String sink) throws PulsarAdminException;
-
-    /**
-     * Create a new sink.
-     *
-     * @param sinkConfig
-     *            the sink configuration object
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void createSink(SinkConfig sinkConfig, String fileName) throws PulsarAdminException;
-
-    /**
-     * <pre>
-     * Create a new sink by providing url from which fun-pkg can be downloaded. supported url: http/file
-     * eg:
-     * File: file:/dir/fileName.jar
-     * Http: http://www.repo.com/fileName.jar
-     * </pre>
-     *
-     * @param sinkConfig
-     *            the sink configuration object
-     * @param pkgUrl
-     *            url from which pkg can be downloaded
-     * @throws PulsarAdminException
-     */
-    void createSinkWithUrl(SinkConfig sinkConfig, String pkgUrl) throws PulsarAdminException;
-
-    /**
-     * Update the configuration for a sink.
-     * <p>
-     *
-     * @param sinkConfig
-     *            the sink configuration object
-     *
-     * @throws NotAuthorizedException
-     *             You don't have admin permission to create the cluster
-     * @throws NotFoundException
-     *             Cluster doesn't exist
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void updateSink(SinkConfig sinkConfig, String fileName) throws PulsarAdminException;
-
-    /**
-     * Update the configuration for a sink.
-     * <p>
-     *
-     * @param sinkConfig
-     *            the sink configuration object
-     * @param updateOptions
-     *            options for the update operations
-     * @throws NotAuthorizedException
-     *             You don't have admin permission to create the cluster
-     * @throws NotFoundException
-     *             Cluster doesn't exist
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void updateSink(SinkConfig sinkConfig, String fileName, UpdateOptions updateOptions) throws PulsarAdminException;
-
-    /**
-     * Update the configuration for a sink.
-     * <pre>
-     * Update a sink by providing url from which fun-pkg can be downloaded. supported url: http/file
-     * eg:
-     * File: file:/dir/fileName.jar
-     * Http: http://www.repo.com/fileName.jar
-     * </pre>
-     *
-     * @param sinkConfig
-     *            the sink configuration object
-     * @param pkgUrl
-     *            url from which pkg can be downloaded
-     * @throws NotAuthorizedException
-     *             You don't have admin permission to create the cluster
-     * @throws NotFoundException
-     *             Cluster doesn't exist
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void updateSinkWithUrl(SinkConfig sinkConfig, String pkgUrl) throws PulsarAdminException;
-
-    /**
-     * Update the configuration for a sink.
-     * <pre>
-     * Update a sink by providing url from which fun-pkg can be downloaded. supported url: http/file
-     * eg:
-     * File: file:/dir/fileName.jar
-     * Http: http://www.repo.com/fileName.jar
-     * </pre>
-     *
-     * @param sinkConfig
-     *            the sink configuration object
-     * @param pkgUrl
-     *            url from which pkg can be downloaded
-     * @param updateOptions
-     *            options for the update operations
-     * @throws NotAuthorizedException
-     *             You don't have admin permission to create the cluster
-     * @throws NotFoundException
-     *             Cluster doesn't exist
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void updateSinkWithUrl(SinkConfig sinkConfig, String pkgUrl, UpdateOptions updateOptions) throws PulsarAdminException;
-
-    /**
-     * Delete an existing sink
-     * <p>
-     * Delete a sink
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @throws NotAuthorizedException
-     *             You don't have admin permission
-     * @throws NotFoundException
-     *             Cluster does not exist
-     * @throws PreconditionFailedException
-     *             Cluster is not empty
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void deleteSink(String tenant, String namespace, String sink) throws PulsarAdminException;
-
-    /**
-     * Gets the current status of a sink.
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    SinkStatus getSinkStatus(String tenant, String namespace, String sink) throws PulsarAdminException;
-
-    /**
-     * Gets the current status of a sink instance.
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     * @param id
-     *            Sink instance-id
-     * @return
-     * @throws PulsarAdminException
-     */
-    SinkStatus.SinkInstanceStatus.SinkInstanceStatusData getSinkStatus(String tenant, String namespace, String sink, int id)
-            throws PulsarAdminException;
-
-    /**
-     * Restart sink instance
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @param instanceId
-     *            Sink instanceId
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void restartSink(String tenant, String namespace, String sink, int instanceId) throws PulsarAdminException;
-
-    /**
-     * Restart all sink instances
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void restartSink(String tenant, String namespace, String sink) throws PulsarAdminException;
-
-
-    /**
-     * Stop sink instance
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @param instanceId
-     *            Sink instanceId
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void stopSink(String tenant, String namespace, String sink, int instanceId) throws PulsarAdminException;
-
-    /**
-     * Stop all sink instances
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void stopSink(String tenant, String namespace, String sink) throws PulsarAdminException;
-
-    /**
-     * Start sink instance
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @param instanceId
-     *            Sink instanceId
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void startSink(String tenant, String namespace, String sink, int instanceId) throws PulsarAdminException;
-
-    /**
-     * Start all sink instances
-     *
-     * @param tenant
-     *            Tenant name
-     * @param namespace
-     *            Namespace name
-     * @param sink
-     *            Sink name
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     */
-    void startSink(String tenant, String namespace, String sink) throws PulsarAdminException;
-
-
-    /**
-     * Fetches a list of supported Pulsar IO sinks currently running in cluster mode
-     *
-     * @throws PulsarAdminException
-     *             Unexpected error
-     *
-     */
-    List<ConnectorDefinition> getBuiltInSinks() throws PulsarAdminException;
 }
