+++ pulsar-functions/worker/src/main/java/org/apache/pulsar/functions/worker/rest/api/v3/SourceApiV3Resource.java
@@ -18,224 +18,20 @@
  */
 package org.apache.pulsar.functions.worker.rest.api.v3;
 
-import io.swagger.annotations.ApiOperation;
-import io.swagger.annotations.ApiResponse;
-import io.swagger.annotations.ApiResponses;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang.StringUtils;
-import org.apache.pulsar.common.functions.UpdateOptions;
-import org.apache.pulsar.common.io.ConnectorDefinition;
-import org.apache.pulsar.common.io.SourceConfig;
-import org.apache.pulsar.common.policies.data.SourceStatus;
-import org.apache.pulsar.functions.worker.rest.FunctionApiResource;
-import org.apache.pulsar.functions.worker.rest.api.SourceImpl;
-import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
-import org.glassfish.jersey.media.multipart.FormDataParam;
+import io.swagger.annotations.Api;
 
-import javax.ws.rs.*;
+import javax.ws.rs.Consumes;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
 
-@Slf4j
 @Path("/source")
-public class SourceApiV3Resource extends FunctionApiResource {
-
-    protected final SourceImpl source;
-
-    public SourceApiV3Resource() {
-        this.source = new SourceImpl(this);
-    }
-
-    @POST
-    @Path("/{tenant}/{namespace}/{sourceName}")
-    @Consumes(MediaType.MULTIPART_FORM_DATA)
-    public void registerSource(final @PathParam("tenant") String tenant,
-                                   final @PathParam("namespace") String namespace,
-                                   final @PathParam("sourceName") String sourceName,
-                                   final @FormDataParam("data") InputStream uploadedInputStream,
-                                   final @FormDataParam("data") FormDataContentDisposition fileDetail,
-                                   final @FormDataParam("url") String functionPkgUrl,
-                                   final @FormDataParam("sourceConfig") String sourceConfigJson) {
-
-        source.registerFunction(tenant, namespace, sourceName, uploadedInputStream, fileDetail,
-                functionPkgUrl, sourceConfigJson, clientAppId(), clientAuthData());
-
-    }
-
-    @PUT
-    @Path("/{tenant}/{namespace}/{sourceName}")
-    @Consumes(MediaType.MULTIPART_FORM_DATA)
-    public void updateSource(final @PathParam("tenant") String tenant,
-                             final @PathParam("namespace") String namespace,
-                             final @PathParam("sourceName") String sourceName,
-                             final @FormDataParam("data") InputStream uploadedInputStream,
-                             final @FormDataParam("data") FormDataContentDisposition fileDetail,
-                             final @FormDataParam("url") String functionPkgUrl,
-                             final @FormDataParam("sourceConfig") String sourceConfigJson,
-                             final @FormDataParam("updateOptions") UpdateOptions updateOptions) {
-
-        source.updateFunction(tenant, namespace, sourceName, uploadedInputStream, fileDetail,
-                functionPkgUrl, sourceConfigJson, clientAppId(), clientAuthData(), updateOptions);
-    }
-
-
-    @DELETE
-    @Path("/{tenant}/{namespace}/{sourceName}")
-    public void deregisterSource(final @PathParam("tenant") String tenant,
-                                 final @PathParam("namespace") String namespace,
-                                 final @PathParam("sourceName") String sourceName) {
-        source.deregisterFunction(tenant, namespace, sourceName, clientAppId(), clientAuthData());
-    }
-
-    @GET
-    @Path("/{tenant}/{namespace}/{sourceName}")
-    public SourceConfig getSourceInfo(final @PathParam("tenant") String tenant,
-                                      final @PathParam("namespace") String namespace,
-                                      final @PathParam("sourceName") String sourceName)
-            throws IOException {
-        return source.getSourceInfo(tenant, namespace, sourceName);
-    }
-
-    @GET
-    @ApiOperation(
-            value = "Displays the status of a Pulsar Source instance",
-            response = SourceStatus.SourceInstanceStatus.SourceInstanceStatusData.class
-    )
-    @ApiResponses(value = {
-            @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 403, message = "The requester doesn't have admin permissions"),
-            @ApiResponse(code = 404, message = "The source doesn't exist")
-    })
-    @Produces(MediaType.APPLICATION_JSON)
-    @Path("/{tenant}/{namespace}/{sourceName}/{instanceId}/status")
-    public SourceStatus.SourceInstanceStatus.SourceInstanceStatusData getSourceInstanceStatus(
-            final @PathParam("tenant") String tenant,
-            final @PathParam("namespace") String namespace,
-            final @PathParam("sourceName") String sourceName,
-            final @PathParam("instanceId") String instanceId) throws IOException {
-        return source.getSourceInstanceStatus(
-            tenant, namespace, sourceName, instanceId, uri.getRequestUri(), clientAppId(), clientAuthData());
-    }
-
-    @GET
-    @ApiOperation(
-            value = "Displays the status of a Pulsar Source running in cluster mode",
-            response = SourceStatus.class
-    )
-    @ApiResponses(value = {
-            @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 403, message = "The requester doesn't have admin permissions"),
-            @ApiResponse(code = 404, message = "The source doesn't exist")
-    })
-    @Produces(MediaType.APPLICATION_JSON)
-    @Path("/{tenant}/{namespace}/{sourceName}/status")
-    public SourceStatus getSourceStatus(final @PathParam("tenant") String tenant,
-                                    final @PathParam("namespace") String namespace,
-                                    final @PathParam("sourceName") String sourceName) throws IOException {
-        return source.getSourceStatus(tenant, namespace, sourceName, uri.getRequestUri(), clientAppId(), clientAuthData());
-    }
-
-    @GET
-    @Path("/{tenant}/{namespace}")
-    public List<String> listSources(final @PathParam("tenant") String tenant,
-                                    final @PathParam("namespace") String namespace) {
-        return source.listFunctions(tenant, namespace, clientAppId(), clientAuthData());
-    }
-
-    @POST
-    @ApiOperation(value = "Restart source instance", response = Void.class)
-    @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 404, message = "The function does not exist"),
-            @ApiResponse(code = 500, message = "Internal server error") })
-    @Path("/{tenant}/{namespace}/{sourceName}/{instanceId}/restart")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public void restartSource(final @PathParam("tenant") String tenant,
-                              final @PathParam("namespace") String namespace,
-                              final @PathParam("sourceName") String sourceName,
-                              final @PathParam("instanceId") String instanceId) {
-        source.restartFunctionInstance(tenant, namespace, sourceName, instanceId, this.uri.getRequestUri(), clientAppId(), clientAuthData());
-    }
-
-    @POST
-    @ApiOperation(value = "Restart all source instances", response = Void.class)
-    @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid request"),
-    @ApiResponse(code = 404, message = "The function does not exist"),
-    @ApiResponse(code = 500, message = "Internal server error") })
-    @Path("/{tenant}/{namespace}/{sourceName}/restart")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public void restartSource(final @PathParam("tenant") String tenant,
-                              final @PathParam("namespace") String namespace,
-                              final @PathParam("sourceName") String sourceName) {
-        source.restartFunctionInstances(tenant, namespace, sourceName, clientAppId(), clientAuthData());
-    }
-
-    @POST
-    @ApiOperation(value = "Stop source instance", response = Void.class)
-    @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 404, message = "The function does not exist"),
-            @ApiResponse(code = 500, message = "Internal server error") })
-    @Path("/{tenant}/{namespace}/{sourceName}/{instanceId}/stop")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public void stopSource(final @PathParam("tenant") String tenant,
-                           final @PathParam("namespace") String namespace,
-                           final @PathParam("sourceName") String sourceName,
-                           final @PathParam("instanceId") String instanceId) {
-        source.stopFunctionInstance(tenant, namespace, sourceName, instanceId, this.uri.getRequestUri(), clientAppId(), clientAuthData());
-    }
-
-    @POST
-    @ApiOperation(value = "Stop all source instances", response = Void.class)
-    @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 404, message = "The function does not exist"),
-            @ApiResponse(code = 500, message = "Internal server error") })
-    @Path("/{tenant}/{namespace}/{sourceName}/stop")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public void stopSource(final @PathParam("tenant") String tenant,
-                           final @PathParam("namespace") String namespace,
-                           final @PathParam("sourceName") String sourceName) {
-        source.stopFunctionInstances(tenant, namespace, sourceName, clientAppId(), clientAuthData());
-    }
-
-    @POST
-    @ApiOperation(value = "Start source instance", response = Void.class)
-    @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 404, message = "The function does not exist"),
-            @ApiResponse(code = 500, message = "Internal server error") })
-    @Path("/{tenant}/{namespace}/{sourceName}/{instanceId}/start")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public void startSource(final @PathParam("tenant") String tenant,
-                            final @PathParam("namespace") String namespace,
-                            final @PathParam("sourceName") String sourceName,
-                            final @PathParam("instanceId") String instanceId) {
-        source.startFunctionInstance(tenant, namespace, sourceName, instanceId, this.uri.getRequestUri(), clientAppId(), clientAuthData());
-    }
-
-    @POST
-    @ApiOperation(value = "Start all source instances", response = Void.class)
-    @ApiResponses(value = { @ApiResponse(code = 400, message = "Invalid request"),
-            @ApiResponse(code = 404, message = "The function does not exist"),
-            @ApiResponse(code = 500, message = "Internal server error") })
-    @Path("/{tenant}/{namespace}/{sourceName}/start")
-    @Consumes(MediaType.APPLICATION_JSON)
-    public void startSource(final @PathParam("tenant") String tenant,
-                            final @PathParam("namespace") String namespace,
-                            final @PathParam("sourceName") String sourceName) {
-        source.startFunctionInstances(tenant, namespace, sourceName, clientAppId(), clientAuthData());
-    }
-
-    @GET
-    @Path("/builtinsources")
-    public List<ConnectorDefinition> getSourceList() {
-        List<ConnectorDefinition> connectorDefinitions = source.getListOfConnectors();
-        List<ConnectorDefinition> retval = new ArrayList<>();
-        for (ConnectorDefinition connectorDefinition : connectorDefinitions) {
-            if (!StringUtils.isEmpty(connectorDefinition.getSourceClass())) {
-                retval.add(connectorDefinition);
-            }
-        }
-        return retval;
-    }
+@Api(value = "/source", description = "Source admin apis", tags = "source")
+@Produces(MediaType.APPLICATION_JSON)
+@Consumes(MediaType.APPLICATION_JSON)
+@Deprecated
+/**
+ * @deprecated in favor of {@link SourcesApiV3Resource}
+ */
+public class SourceApiV3Resource extends SourcesApiV3Resource {
 }
